+++ react/components/ProductSpecifications/index.js
@@ -122,12 +122,26 @@ const ProductSpecifications = ({
         {specificationItems.map((specification, i) => (
           <tr key={i}>
             <td
-              className={`${handles.specificationItemProperty} w-50 b--muted-4 bb pa5`}
+              className={`${
+                handles.specificationItemProperty
+              } ${handles.specificationItemProperty +
+                '-' +
+                specification.property.replace(
+                  /\W+/g,
+                  '_'
+                )} w-50 b--muted-4 bb pa5`}
             >
               {HtmlParser(specification.property)}
             </td>
             <td
-              className={`${handles.specificationItemSpecifications} w-50 b--muted-4 bb pa5`}
+              className={`${
+                handles.specificationItemSpecifications
+              } ${handles.specificationItemProperty +
+                '-' +
+                specification.specifications.replace(
+                  /\W+/g,
+                  '_'
+                )} w-50 b--muted-4 bb pa5`}
             >
               {HtmlParser(specification.specifications)}
             </td>