+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed Product Price block documentation by adding a deprecation badge and disclaimer.
 
+### Added
+- New CSS handle `specificationItemProperty` plus the name of the specification.
+
 ## [3.115.0] - 2020-05-20
 ### Added
 - Option to edit `callToActionLinkTarget` prop from the InfoCard on site editor.